total.go
