+++ server/total.go
@@ -1,57 +1,68 @@
 package server
 
 import (
+	"encoding/json"
 	"hscan/models"
-	"net/http"
 
 	"github.com/gin-gonic/gin"
 	"github.com/pkg/errors"
 )
+
+func sortCoins(Coins []map[string]interface{}) []map[string]interface{} {
+
+	for i := 0; i < len(Coins); {
+
+		if Coins[i]["denom"] == "syscoin" || Coins[i]["denom"] == "SYSCOIN" {
+			Coins = append(Coins[:i], Coins[i+1:]...)
+
+			continue
+		}
+
+		if Coins[i]["denom"] == "hst" || Coins[i]["denom"] == "uhst" {
+			a := Coins[i]
+			Coins[i] = Coins[0]
+			Coins[0] = a
+		}
+		i++
+	}
+
+	if len(Coins) == 0 {
+		hst := make(map[string]interface{}, 1)
+		hst["amount"] = "0"
+		hst["denom"] = "uhst"
+		Coins = make([]map[string]interface{}, 1)
+		Coins[0] = hst
+		return Coins
+	}
+
+	if Coins[0]["denom"] != "uhst" {
+		hst := make([]map[string]interface{}, 1)
+		hst[0]["amount"] = "0"
+		hst[0]["denom"] = "uhst"
+		Coins = append(hst, Coins...)
+		return Coins
+	}
+	return Coins
+}
 
 func (s *Server) totals(c *gin.Context) {
 
 	status, err := s.client.Querytotals()
 	if err != nil {
 		s.l.Print(errors.Wrap(err, "failed to query the latest block height on the active network"))
-		status = nil
-		c.JSON(http.StatusOK, gin.H{
-			"data": nil,
-		})
+		s.interfaceResponse(c, nil)
 		return
 	}
 
-	body, _ := s.parseResponse(status)
-
-	for i := 0; i < len(body["result"].([]interface{})); {
-		denom := body["result"].([]interface{})[i].(map[string]interface{})["denom"]
-
-		if denom == "syscoin" || denom == "SYSCOIN" {
-			body["result"] = append(body["result"].([]interface{})[:i], body["result"].([]interface{})[i+1:]...)
-
-			continue
-		}
-
-		if Priceinto, OK := s.Priceinto[denom.(string)]; OK {
-			body["result"].([]interface{})[i].(map[string]interface{})["price"] = Priceinto.Pirce
-			body["result"].([]interface{})[i].(map[string]interface{})["priceunit"] = Priceinto.Priceunit
-
-		} else {
-			var Priceinto models.Priceinto
-			num, priceunit, err := s.getDenomPrice(denom)
-			body["result"].([]interface{})[i].(map[string]interface{})["price"] = num.(string)
-			body["result"].([]interface{})[i].(map[string]interface{})["priceunit"] = priceunit.(string)
-			if err == nil {
-				Priceinto.Pirce = num.(string)
-				Priceinto.Priceunit = priceunit.(string)
-				s.Priceinto[denom.(string)] = Priceinto
-			}
-
-		}
-		i++
+	var result models.TotalInfo
+	err = json.Unmarshal(status.Body(), &result)
+	if err != nil {
+		s.interfaceResponse(c, nil)
+		return
 	}
-	c.JSON(http.StatusOK, gin.H{
-		"data": body,
-	})
+
+	result.Result, _ = s.CoinsPrice(result.Result)
+	s.interfaceResponse(c, result)
 }
 
 func (s *Server) total(c *gin.Context) {
@@ -60,32 +71,12 @@ func (s *Server) total(c *gin.Context) {
 	parameters, err := s.client.Querytotal(denomination)
 	if err != nil {
 		s.l.Print(errors.Wrap(err, "failed to query the latest block height on the active network"))
-		parameters = nil
-		c.JSON(http.StatusOK, gin.H{
-			"data": nil,
-		})
+		s.interfaceResponse(c, nil)
 		return
 	}
+
 	body, _ := s.parseResponse(parameters)
+	body, _ = s.denomPrice(body, denomination)
 
-	if Priceinto, OK := s.Priceinto[denomination]; OK {
-		body["price"] = Priceinto.Pirce
-		body["priceunit"] = Priceinto.Priceunit
-
-	} else {
-		var Priceinto models.Priceinto
-		num, priceunit, err := s.getDenomPrice(denomination)
-		body["price"] = num.(string)
-		body["priceunit"] = priceunit.(string)
-		if err == nil {
-			Priceinto.Pirce = num.(string)
-			Priceinto.Priceunit = priceunit.(string)
-			s.Priceinto[denomination] = Priceinto
-		}
-
-	}
-
-	c.JSON(http.StatusOK, gin.H{
-		"data": body,
-	})
+	s.interfaceResponse(c, body)
 }
